scripts/media/requirements.txt
# Photo Consolidation Python Dependencies
pyyaml>=6.0.1
tqdm>=4.65.0
pillow>=10.0.0
exifread>=3.0.0
click>=8.1.0
colorama>=0.4.6
pathlib2>=2.3.7
psutil>=5.9.0
pytest>=7.4.0